chore(cleanup): Remove unneeded files
There were a few other files in this repository that weren't strictly
related to the needs of the cards. While this gives the repository a
little extra flair, I'm trying to keep this hyper focused for the set of
Cards and ensure that it's appropriate and relevant for my team.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>